Move tests to bulk_subscriber_test file
